@@ -23,52 +23,70 @@
   ],
   "author": "jerryshew",
   "license": "MIT",
-  "homepage": "https://github.com",
+  "homepage": "https://github.com/jerryshew",
   "dependencies": {
-    "deep-extend": "^0.4.1",
-    "npm-run-all": "^4.1.1",
-    "prop-types": "^15.6.0",
-    "react": "^16.2.0",
-    "react-dom": "^16.2.0"
+    "deep-extend": "^0.6.0",
+    "npm-run-all": "^4.1.5",
+    "prop-types": "15.6.0",
+    "react": "16.5.2",
+    "react-dom": "16.5.2"
   },
   "pre-commit": [
     "lint"
   ],
   "devDependencies": {
-    "autoprefixer": "^7.1.1",
-    "babel-cli": "^6.18.0",
-    "babel-core": "^6.18.2",
-    "babel-loader": "^6.0.0",
-    "babel-plugin-react-transform": "^1.1.1",
-    "babel-polyfill": "^6.23.0",
-    "babel-preset-es2015": "^6.6.0",
-    "babel-preset-react": "^6.5.0",
-    "babel-runtime": "^6.0.0",
-    "cross-env": "^3.1.3",
-    "css-loader": "^0.28.4",
-    "eslint": "^4.16.0",
-    "eslint-config-standard": "^11.0.0-beta.0",
-    "eslint-plugin-import": "^2.8.0",
-    "eslint-plugin-node": "^5.2.1",
-    "eslint-plugin-promise": "^3.6.0",
-    "eslint-plugin-react": "^7.6.1",
-    "eslint-plugin-standard": "^3.0.1",
-    "express": "^4.13.3",
-    "extract-text-webpack-plugin": "^2.1.2",
-    "file-loader": "^0.8.5",
-    "force-case-sensitivity-webpack-plugin": "^0.1.1",
-    "html-webpack-plugin": "^1.6.1",
-    "less": "^2.7.2",
-    "less-loader": "^4.0.5",
-    "postcss-loader": "^2.0.6",
+    "@babel/cli": "^7.2.0",
+    "@babel/core": "^7.2.0",
+    "@babel/plugin-proposal-class-properties": "^7.2.1",
+    "@babel/plugin-syntax-dynamic-import": "^7.2.0",
+    "@babel/preset-env": "^7.2.0",
+    "@babel/preset-react": "^7.0.0",
+    "autoprefixer-loader": "^3.1.0",
+    "babel-eslint": "^10.0.1",
+    "babel-loader": "^8.0.4",
+    "babel-plugin-dynamic-import-node": "^2.2.0",
+    "babel-polyfill": "^6.26.0",
+    "babel-runtime": "^6.26.0",
+    "cross-env": "^1.0.8",
+    "css-loader": "^0.28.11",
+    "eslint": "^5.9.0",
+    "eslint-config-prettier": "^3.3.0",
+    "eslint-config-standard": "^12.0.0",
+    "eslint-config-standard-react": "^7.0.2",
+    "eslint-plugin-import": "^2.14.0",
+    "eslint-plugin-jsx-a11y": "^6.1.2",
+    "eslint-plugin-node": "^8.0.0",
+    "eslint-plugin-prettier": "^3.0.0",
+    "eslint-plugin-promise": "^4.0.1",
+    "eslint-plugin-react": "^7.11.1",
+    "eslint-plugin-standard": "^4.0.0",
+    "express": "^4.16.4",
+    "file-loader": "^1.1.11",
+    "koa-connect": "^2.0.1",
+    "less": "^3.9.0",
+    "less-loader": "^4.1.0",
+    "mini-css-extract-plugin": "^0.4.4",
+    "node-sass": "^4.11.0",
+    "postcss-loader": "^2.1.6",
     "pre-commit": "^1.2.2",
-    "progress-bar-webpack-plugin": "^1.9.3",
+    "prettier": "^1.15.3",
+    "prettier-standard": "^8.0.1",
     "react-transform-hmr": "^1.0.0",
-    "style-loader": "^0.12.3",
-    "uglify-loader": "^1.2.0",
-    "url-loader": "^0.5.7",
-    "webpack": "^2.6.1",
-    "webpack-dev-middleware": "^1.2.0",
-    "webpack-hot-middleware": "^2.5.0"
+    "sass-loader": "^6.0.7",
+    "style-loader": "^0.21.0",
+    "uglifyjs-webpack-plugin": "^1.3.0",
+    "webpack": "^4.20.2",
+    "webpack-bundle-analyzer": "^3.0.2",
+    "webpack-cli": "^3.1.2",
+    "webpack-hot-client": "^4.1.1",
+    "webpack-merge": "^4.1.4",
+    "webpack-serve": "^1.0.4"
+  },
+  "standard": {
+    "parser": "babel-eslint",
+    "ignore": [
+      "dist/",
+      "build/"
+    ]
   }
 }
